feat(runtime)!: replace channels with observers
This commit replaces channels in favor of the observer design pattern. This model allows the server to send messages to the client as soon as a message is ready for sending. This removes the need for the client to be constantly polling itself, so that it could process newly arrived messages.

BREAKING CHANGE: all channel tests, classes, and methods are deleted